update amount of balance from int to float

+++ internal/app/balance/delivery/models/response_models.go
@@ -5,12 +5,12 @@ type ErrResponse struct {
 }
 
 type ResponseBalance struct {
-	UserID  int64 `json:"user_id,omitempty"`
-	Balance int64 `json:"balance"`
+	UserID  int64   `json:"user_id,omitempty"`
+	Balance float64 `json:"balance"`
 }
 type ResponseTransfer struct {
-	SenderID        int64 `json:"sender_id"`
-	SenderBalance   int64 `json:"sender_balance"`
-	ReceiverID      int64 `json:"receiver_id"`
-	ReceiverBalance int64 `json:"receiver_balance"`
+	SenderID        int64   `json:"sender_id"`
+	SenderBalance   float64 `json:"sender_balance"`
+	ReceiverID      int64   `json:"receiver_id"`
+	ReceiverBalance float64 `json:"receiver_balance"`
 }
